Create Data input- with add function
this is the code in which when the user enters the number of data inputs, the number of input boxes are created and in them he enters the data. after this for now I added an option for addition but we can add more options such as the mathematical calculations.